添加左右分页的日间模式和黑夜模式，以及字体调节功能，以及目录跳转等功能

+++ lib/views/CoustomSlider.dart
@@ -29,7 +29,7 @@ class CoustomSliderState extends State<CoustomSlider>{
         fontsize = value.toInt();
         Tools.updateUI(this);
         if(state is ChapterDetailState){
-          (state as ChapterDetailState).fontsize = fontsize;
+          (state as ChapterDetailState).fontsize = fontsize.toDouble();
           (state as ChapterDetailState).updateFontSize();
         }
       },
